% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxreg_bias.R
\name{coxreg_bias}
\alias{coxreg_bias}
\title{bias analysis for cox regression}
\usage{
coxreg_bias(data, surv, main, bnry = NULL, real = NULL,
  xtra.adj = NULL, bnry.manual = NULL, real.manual = NULL,
  rms = FALSE, regfnc = NULL, verbose = TRUE)
}
\arguments{
\item{data}{data frame}

\item{surv}{name of response 'Surv'-variable in data set}

\item{main}{name of main effect (binary)}

\item{bnry}{names of binary variables to include}

\item{real}{names of continuous 'normalish' variables to include}

\item{xtra.adj}{extra string to include in formula}

\item{bnry.manual}{additional binary  confounders not in model,  named list
where each element is HR, proportion  at \code{main == 0}, proportion at
\code{main == 1}}

\item{real.manual}{additional real  confounders not  in model,  named list
where each element  is HR, mean at \code{main ==  0}, mean at \code{main
== 1}}

\item{rms}{use \code{rms::cph} instead of \code{survival::coxph}? (mainly if
xtra.adj contains something specific to the rms package)}

\item{regfnc}{a regression function taking a formula- and data
argument. Experimental, probably must be a wrapper around \code{coxph}
in order to work.}

\item{verbose}{logical; send a message if deemed necessary?}
}
\description{
use obsSens package to get estimates  of 'true' effect given the addition of
    a confounder
}
