% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxreg_bias.R
\name{coxreg_bias2plot}
\alias{coxreg_bias2plot}
\title{coxreg bias plot data}
\usage{
coxreg_bias2plot(x)
}
\arguments{
\item{x}{the return of a \code{coxreg_bias} call}
}
\value{
a data frame
}
\description{
helper function to get (what I deem) the important parts of the results of
    \code{coxreg_bias} into a tidy format suitable for plotting
}
