% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxreg_bias.R
\name{updHRbnry}
\alias{updHRbnry}
\alias{updHRnorm}
\title{bias adjusted HR}
\usage{
updHRbnry(HR, expG, p0, p1)

updHRnorm(HR, expG, m0, m1)
}
\arguments{
\item{HR}{the treatment HR (to be updated) on the outcome}

\item{expG}{the confounder HR on the outcome}

\item{p0}{proportion among controls}

\item{p1}{proportion among cases}

\item{m0}{mean value among controls}

\item{m1}{mean value among cases}
}
\description{
calculate what  treatment HR would be  if we could adjust  for an unmeasured
    binary confounder  U having  proportion p0  and p1  in the  control- and
    treatment group, respectively, and whose HR  on the outcome is expG. The
    formulas   are   from  Lin,   D.   Y.,   Psaty,   B.  M.,   &   Kronmal,
    R.  A.  (1998).  Assessing  the sensitivity  of  regression  results  to
    unmeasured confounders in observational studies. Biometrics, 948-963.
}
\section{Functions}{
\itemize{
\item \code{updHRnorm}: function for when U is normal (instead of binary)
}}

