% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxreg_HRS.R
\name{coxreg_HRS}
\alias{coxreg_HRS}
\title{treatment HR changes}
\usage{
coxreg_HRS(data, surv, main, terms, decr.inc = NULL, decr.exc = NULL,
  rms = FALSE)
}
\arguments{
\item{data}{data frame}

\item{surv}{name of response 'Surv'-variable in data set}

\item{main}{name of main effect (binary)}

\item{terms}{vector of terms to adjust for (can be named)}

\item{decr.inc}{logical; decreasing order for sequential inclusion?}

\item{decr.exc}{logical; decreasing order for sequential exclusion?}

\item{rms}{use \code{rms::cph} instead of \code{survival::coxph}? (mainly if
xtra.adj contains something specific to the rms package)}
}
\description{
examine changes in main effect (HR, Cox regression) when covariates are
    added one at a time, sequentially increasing (starting from a small
    model) or sequentially decreasing (starting from large model)
}
